+++ app-frontend/src/app/components/pages/project/settings/index.js
@@ -2,9 +2,11 @@ import optionsPage from './options';
 import permissionsPage from './permissions';
 import publishingPage from './publishing';
 import settingsPage from './settings';
+import embedPage from './embed';
 export default [
     optionsPage,
     permissionsPage,
     publishingPage,
-    settingsPage
+    settingsPage,
+    embedPage
 ];