+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Added project settings pages for v2 UI [\#4637](https://github.com/raster-foundry/raster-foundry/pull/4637)
 - Added project layer color-mode UI [\#4706](https://github.com/raster-foundry/raster-foundry/pull/4706)
 - Added project layer color-correction UI [\#4722](https://github.com/raster-foundry/raster-foundry/pull/4722)
+- Added project embed UI [\#4793](https://github.com/raster-foundry/raster-foundry/pull/4793/files)
 - Added single-band color-mode support to project layer color-mode UI [\#4728](https://github.com/raster-foundry/raster-foundry/pull/4728)
 - Added owner query parameter to tools and tool-runs endpoints, support multiple owner qp's on applicable endpoints [\#4689](https://github.com/raster-foundry/raster-foundry/pull/4689)
 - Added Rollbar error reporting to backsplash [\#4691](https://github.com/raster-foundry/raster-foundry/pull/4691)